+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.3.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.3.0...@stream-io/video-react-native-sdk-0.3.1) (2023-11-30)
+
+
+### Bug Fixes
+
+* **react-native:** do not trigger initial device setting after join ([204c303](https://github.com/GetStream/stream-video-js/commit/204c303353c536c44b77350bb49c117f21e093c5))
+
+## [0.3.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.2.14...@stream-io/video-react-native-sdk-0.3.0) (2023-11-29)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.5.0`
+* `@stream-io/video-react-bindings` updated to version `0.3.11`
+
+### ⚠ BREAKING CHANGES
+
+* **react-native:** move to webrtc 118 (#1197)
+
+### Features
+
+* **react-native:** move to webrtc 118 ([#1197](https://github.com/GetStream/stream-video-js/issues/1197)) ([8cdbe11](https://github.com/GetStream/stream-video-js/commit/8cdbe11de069fcb6eae5643f5cef5c9612f6c805))
+
+
+### Bug Fixes
+
+* **react-native:** remove unused import ([388d5fc](https://github.com/GetStream/stream-video-js/commit/388d5fc41479190c34b1f5042303157b96149381))
+
 ### [0.2.14](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.2.13...@stream-io/video-react-native-sdk-0.2.14) (2023-11-27)
 
 ### Dependency Updates